pyproject.toml
[project]
name = "agentic-astra"
version = "0.0.5"
description = "Astra MCP Server"
readme = "README.md"
requires-python = ">=3.12"

authors = [
    {name = "Samuel Matioli", email = "smatioli@gmail.com"},
    {name = "Samuel Matioli", email = "samuel.matioli@ibm.com"}
]

license = {file = "LICENSE"}
keywords = ["mcp", "astra", "datastax", "server"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "astrapy>=2.0.1",
    "fastmcp>=2.12.1",
    "python-dotenv>=1.1.1",
    "uvicorn[standard]>=0.30.0",
]

[project.urls]
Homepage = "https://github.com/smatiolids/agentic-astra"
Repository = "https://github.com/smatiolids/agentic-astra"
Issues = "https://github.com/smatiolids/agentic-astra/issues"

[tool.uv]
dev-dependencies = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "build>=1.3.0",
    "twine>=6.2.0",
]

[project.scripts]
agentic-astra = "agentic_astra.server:run_server"
agentic-astra-catalog = "agentic_astra.catalog:main"
agentic-astra-tool-agent = "agentic_astra.tool_agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentic_astra"]